requirements.txt
Django==3.0.10
djangorestframework==3.11.1
Pillow==7.2.0


